Move test for dll wrapper into it's own test module
